refactor: createdAt 타입 변경
